ArithmeticCalculator 클래스에 추가로 나머지 연산(%) 기능을 추가하기 위해 ModOperator 클래스를 만들어 추가합니다.

+++ src/Task2/AbstractOperation.java
@@ -1,5 +1,5 @@
 package Task2;
 
 public abstract class AbstractOperation {
-    abstract double operate(int firstNumber, int secondNumber);
+    abstract double operate(double firstNumber, double secondNumber);
 }
